Moving config and db module to utils
