+++ backend/app/utils/db.py
@@ -4,7 +4,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import as_declarative
 from sqlalchemy.orm import sessionmaker
 
-from .config import settings
+from app.utils.config import settings
 
 engine = create_engine(
     settings.SQLALCHEMY_DATABASE_URI,
